feat: prompts
Remove old hackjob/copypaste password prompting system. Update to prompts library. Add confirmation steps for several commands (remove, decrypt).

+++ src/commands/remove.js
@@ -1,6 +1,19 @@
 const { noArraysExcept } = require('../utils');
-const { deleteSecretByAlias } = require('../db');
+const { deleteSecretByAlias, cleanup } = require('../db');
 const printer = require('../printer');
+const prompts = require('prompts');
+
+const confirmationPrompt = {
+  type: 'confirm',
+  name: 'confirmed',
+  message:
+    'REMOVING A SECRET CANNOT BE UNDONE. Are you sure you want to remove the specified secret?',
+};
+
+const onCancel = () => {
+  cleanup(true);
+  process.exit(0);
+};
 
 module.exports = {
   command: ['remove <alias>', 'delete', 'rm'],
@@ -15,8 +28,13 @@ module.exports = {
       .check(noArraysExcept([]), false)
       .group(['alias'], 'REMOVE options:');
   },
-  handler: (options) => {
+  handler: async (options) => {
     const { alias, verbose } = options;
+    const { confirmed } = await prompts(confirmationPrompt, { onCancel });
+    if (!confirmed) {
+      if (verbose) console.log(printer.verbose('User did not confirm action.'));
+      return;
+    }
     if (verbose)
       console.log(printer.verbose(`Deleting secret by alias: ${alias}`));
     const success = deleteSecretByAlias(alias);
